Renamed programInput to TextInput
